# public/openapi.yaml
openapi: 3.0.3
info:
  title: E-Puskesmas EIS API
  version: 1.0.0
  description: |
    OpenAPI untuk E-Puskesmas EIS. Disusun mengikuti struktur menu frontend:
    Dashboard, Cari, Klaster 1-4, Lintas Klaster, Monitoring, dan Akun.

    **Integrasi pihak ketiga (e-Puskesmas / webhook)**
    - Semua pengiriman data dari sistem eksternal dilakukan via endpoint `POST /api/eis/klaster1`, `POST /api/eis/klaster2`, `POST /api/eis/klaster3`, `POST /api/eis/klaster4`, dan `POST /api/eis/lintas-klaster`.
    - Endpoint dashboard/analitik tetap `GET` (read-only). Data yang tampil berasal dari hasil POST (webhook push).
    - Semua endpoint POST di atas wajib menyertakan header `X-Epus-Signature` untuk HMAC:
      signature = HMAC_SHA256(EPUS_WEBHOOK_SECRET, `${t}.${rawJsonUtf8}`)
      Header format: `t=<unix_ts>,v1=<hex_sha256>`
    - Idempotency: gunakan `batchId` unik agar retry aman (tidak dobel).
    - Skala besar: payload boleh `Content-Encoding: gzip` (signature dihitung dari JSON *setelah* di-decompress).
servers:
  - url: http://localhost:3000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  parameters:
    XEpusSignature:
      name: X-Epus-Signature
      in: header
      required: true
      description: "Signature HMAC. Format: `t=<unix_ts>,v1=<hex_sha256>`"
      schema:
        type: string
        example: t=1700000000,v1=0123abcd...
    ContentEncoding:
      name: Content-Encoding
      in: header
      required: false
      description: "Opsional untuk payload besar. Gunakan `gzip` bila body di-compress."
      schema:
        type: string
        enum: [gzip]
...
    PuskesmasId:
      name: puskesmasId
      in: query
      required: false
      schema:
        type: string
      description: "ID puskesmas, gunakan `all` untuk semua"
    Bulan:
      name: bulan
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 12
      description: Bulan (1-12)
    Tahun:
      name: tahun
      in: query
      required: false
      schema:
        type: integer
      description: Tahun (YYYY)
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    KlasterGenericResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          additionalProperties: true
...
paths:
...
  /api/eis/klaster1:
    get:
      tags: [Klaster 1]
      summary: Data Klaster 1
      parameters:
        - $ref: '#/components/parameters/PuskesmasId'
        - $ref: '#/components/parameters/Bulan'
        - $ref: '#/components/parameters/Tahun'
      responses:
        '200':
          description: Data klaster 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KlasterGenericResponse'
    post:
      tags: [Klaster 1]
      summary: Webhook push data Klaster 1 (SDM, Obat, Keuangan)
      description: |
        Dipakai oleh sistem eksternal (e-Puskesmas) untuk mengirim **data transaksi/agregat bulanan** Klaster 1.

        Minimal kirim salah satu: `tenagaKesehatan`, `stokObat`, `keuangan`.
        Server akan **upsert** berdasarkan kombinasi unik:
        - Tenaga Kesehatan: `puskesmas + kategori + bulan + tahun`
        - Stok Obat: `puskesmas + namaObat + bulan + tahun`
        - Keuangan: `puskesmas + kategori + bulan + tahun`
      parameters:
        - $ref: '#/components/parameters/XEpusSignature'
        - $ref: '#/components/parameters/ContentEncoding'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                batchId: { type: string }
                puskesmasKode: { type: string }
                puskesmasId: { type: string }
                bulan: { type: integer, minimum: 1, maximum: 12 }
                tahun: { type: integer }
                tenagaKesehatan:
                  type: array
                  items:
                    type: object
                    required: [kategori, jumlah, target]
                    properties:
                      kategori: { type: string }
                      jumlah: { type: integer }
                      target: { type: integer }
                stokObat:
                  type: array
                  items:
                    type: object
                    required: [namaObat, satuan, stok, pemakaian]
                    properties:
                      namaObat: { type: string }
                      satuan: { type: string }
                      stok: { type: integer }
                      pemakaian: { type: integer }
                keuangan:
                  type: array
                  items:
                    type: object
                    required: [kategori, nominal]
                    properties:
                      kategori: { type: string }
                      nominal: { type: number }
              oneOf:
                - required: [batchId, puskesmasKode, bulan, tahun]
                - required: [batchId, puskesmasId, bulan, tahun]
            examples:
              full_sdm_obat_keuangan:
                summary: Kirim SDM + Obat + Keuangan (lengkap)
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K1-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  tenagaKesehatan:
                    - kategori: Dokter Umum
                      jumlah: 2
                      target: 3
                    - kategori: Dokter Gigi
                      jumlah: 1
                      target: 1
                    - kategori: Bidan
                      jumlah: 6
                      target: 7
                  stokObat:
                    - namaObat: Paracetamol 500mg
                      satuan: tablet
                      stok: 2500
                      pemakaian: 800
                    - namaObat: Amoxicillin 500mg
                      satuan: kapsul
                      stok: 1200
                      pemakaian: 350
                  keuangan:
                    - kategori: Pendapatan JKN
                      nominal: 12500000
                    - kategori: Pendapatan Umum
                      nominal: 3500000
              only_obat:
                summary: Kirim Obat saja (boleh tanpa SDM & Keuangan)
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K1-OBAT-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  stokObat:
                    - namaObat: Paracetamol 500mg
                      satuan: tablet
                      stok: 2500
                      pemakaian: 800
              only_sdm:
                summary: Kirim SDM saja
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K1-SDM-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  tenagaKesehatan:
                    - kategori: Dokter Umum
                      jumlah: 2
                      target: 3
              only_keuangan:
                summary: Kirim Keuangan saja
                value:
                  batchId: EPUS-BUMIAYU-2026-02-K1-KEU-0001
                  puskesmasKode: "3329XXXX"
                  bulan: 2
                  tahun: 2026
                  keuangan:
                    - kategori: Pendapatan JKN
                      nominal: 12500000
      responses:
        '200':
          description: Data diterima dan diproses
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '400':
          description: Validasi gagal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Signature invalid
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Kesalahan server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...